pass torch dtype through from_pretrained.

+++ llava/model/multimodal_encoder/clip_encoder.py
@@ -1,4 +1,6 @@
 # This file is modified from https://github.com/haotian-liu/LLaVA/
+import torch
+
 from llava.model.multimodal_encoder.vision_encoder import VisionTower
 from transformers import (
     PretrainedConfig,
@@ -11,5 +13,7 @@ class CLIPVisionTower(VisionTower):
     def __init__(self, model_name_or_path: str, config: PretrainedConfig):
         super().__init__(model_name_or_path, config)
         self.image_processor = CLIPImageProcessor.from_pretrained(model_name_or_path)
-        self.vision_tower = CLIPVisionModel.from_pretrained(model_name_or_path)
+        self.vision_tower = CLIPVisionModel.from_pretrained(
+            model_name_or_path, dtype=torch.bfloat16
+        )
         self.is_loaded = True
