+++ llava/model/multimodal_encoder/siglip_encoder.py
@@ -1,3 +1,4 @@
+import torch
 from llava.model.multimodal_encoder.vision_encoder import VisionTower
 
 from transformers import PretrainedConfig
@@ -11,5 +12,7 @@ class SiglipVisionTower(VisionTower):
     def __init__(self, model_name_or_path: str, config: PretrainedConfig):
         super().__init__(model_name_or_path, config)
         self.image_processor = SiglipImageProcessor.from_pretrained(model_name_or_path)
-        self.vision_tower = SiglipVisionModel.from_pretrained(model_name_or_path)
+        self.vision_tower = SiglipVisionModel.from_pretrained(
+            model_name_or_path, dtype=torch.bfloat16
+        )
         self.is_loaded = True
